+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.12.0] - 2026-03-11
+
+- Upgrade to SQLite3 Multiple Ciphers version 2.3.0 (SQLite version 3.52.0)
+- Add build support for VS2026
+- Unify architecture names (use _x64_ instead of _Win64_) to avoid conflicts with VS2026
+
 ## [4.11.2] - 2026-01-12
 
 - Upgrade to SQLite3 Multiple Ciphers version 2.2.7 (SQLite version 3.51.2)
@@ -717,7 +723,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release
 
-[Unreleased]: ../../compare/v4.11.2...HEAD
+[Unreleased]: ../../compare/v4.12.0...HEAD
+[4.12.0]: ../../compare/v4.11.2...v4.12.0
 [4.11.2]: ../../compare/v4.11.1...v4.11.2
 [4.11.1]: ../../compare/v4.11.0...v4.11.1
 [4.11.0]: ../../compare/v4.10.12...v4.11.0